pyproject.toml
[build-system]
requires = [
    "setuptools>=68",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "vla-scratch"
version = "0.1.0"
description = "Utilities for working with VLA Scratch datasets and policies."
requires-python = "==3.10.*"
dependencies = [
    "torch",
    "torchvision==0.22.0",
    "numpy",
    "h5py",
    "transformers==4.57.1",
    "jaxtyping",
    "fsspec",
    "filelock",
    "tqdm",
    "einops",
    "tensordict",
    "hydra-core",
    "omegaconf",
    "wandb",
    "setproctitle",
    "pillow",
    "setuptools",
    # serve policy
    "pyzmq",
    "msgpack",
    "pdbpp",
    "datasets",
    "jsonlines",
    "matplotlib",
    # the last commit before dataset version 3.0
    "lerobot @ git+https://github.com/huggingface/lerobot@d602e8169cbad9e93a4a3b3ee1dd8b332af7ebf8",
    "torchcodec",
    "accelerate",
    "mujoco",
    "viser",
    # "flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu12torch2.7cxx11abiTRUE-cp310-cp310-linux_x86_64.whl",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "vla_scratch",
    "vla_scratch.*",
]
namespaces = true

[tool.uv]
python-install-mirror = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
